requirements.txt
SAMITorch>=0.1.32
pytorch-ignite>=0.2.0
PyYAML>=5.1
pytorch-ignite>=0.2.0